# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "polyGraphics 3D Backend",
    "description": "API for 2D-to-3D reconstruction. Default path: **Segment Anything first** (smart background removal), then MapAnything / mesh or Gaussian splatting \u2014 reconstructions consume **masked** views only. `skip_sam_segmentation` is only for pre-cut **RGBA** assets with real transparency.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Direct Uvicorn (e.g. http://127.0.0.1:8000)"
    }
  ],
  "paths": {
...
    "/isolation/quota": {
      "get": {
        "tags": [
          "isolation"
        ],
        "summary": "Isolation Quota",
        "operationId": "isolation_quota_isolation_quota_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationQuotaStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/isolation/datasets": {
      "get": {
        "tags": [
          "isolation"
        ],
        "summary": "List Datasets",
        "operationId": "list_datasets_isolation_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IsolationDatasetSummary"
                  },
                  "type": "array",
                  "title": "Response List Datasets Isolation Datasets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "isolation"
        ],
        "summary": "Create Dataset",
        "operationId": "create_dataset_isolation_datasets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsolationDatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationDatasetSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/isolation/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "isolation"
        ],
        "summary": "Get Dataset",
        "operationId": "get_dataset_isolation_datasets__dataset_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationDatasetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "isolation"
        ],
        "summary": "Delete Dataset",
        "operationId": "delete_dataset_isolation_datasets__dataset_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/isolation/datasets/{dataset_id}/pairs": {
      "post": {
        "tags": [
          "isolation"
        ],
        "summary": "Upload Pairs",
        "operationId": "upload_pairs_isolation_datasets__dataset_id__pairs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_pairs_isolation_datasets__dataset_id__pairs_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationPairUploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/isolation/train": {
      "post": {
        "tags": [
          "isolation"
        ],
        "summary": "Start Train",
        "operationId": "start_train_isolation_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsolationTrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationTrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/isolation/train/{job_id}": {
      "get": {
        "tags": [
          "isolation"
        ],
        "summary": "Get Train",
        "operationId": "get_train_isolation_train__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationTrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/isolation/models": {
      "get": {
        "tags": [
          "isolation"
        ],
        "summary": "List Models",
        "operationId": "list_models_isolation_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IsolationModelSummary"
                  },
                  "type": "array",
                  "title": "Response List Models Isolation Models Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/isolation/models/active": {
      "get": {
        "tags": [
          "isolation"
        ],
        "summary": "Get Active Model",
        "operationId": "get_active_model_isolation_models_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationModelSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/isolation/models/{model_id}/activate": {
      "post": {
        "tags": [
          "isolation"
        ],
        "summary": "Activate Model",
        "operationId": "activate_model_isolation_models__model_id__activate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationModelSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/isolation/models/import": {
      "post": {
        "tags": [
          "isolation"
        ],
        "summary": "Import Model",
        "operationId": "import_model_isolation_models_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_model_isolation_models_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsolationModelSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/isolation/predict": {
      "post": {
        "tags": [
          "isolation"
        ],
        "summary": "Predict Isolation",
        "operationId": "predict_isolation_isolation_predict_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "png or json",
              "default": "png",
              "title": "Format"
            },
            "description": "png or json"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_predict_isolation_isolation_predict_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_model_isolation_models_import_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "model.onnx"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "activate": {
            "type": "boolean",
            "title": "Activate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "file"
        ],
        "title": "Body_import_model_isolation_models_import_post"
      },
...
      "Body_predict_isolation_isolation_predict_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_predict_isolation_isolation_predict_post"
      },
...
      "Body_upload_pairs_isolation_datasets__dataset_id__pairs_post": {
        "properties": {
          "indices": {
            "type": "string",
            "title": "Indices",
            "description": "Comma-separated pair indices, e.g. 0,1,2"
          },
          "before": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Before",
            "description": "Before images (same order as indices)"
          },
          "after": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "After",
            "description": "After studio cutout images"
          },
          "mask": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mask",
            "description": "Optional explicit masks"
          }
        },
        "type": "object",
        "required": [
          "indices",
          "before",
          "after"
        ],
        "title": "Body_upload_pairs_isolation_datasets__dataset_id__pairs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IsolationDatasetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "IsolationDatasetCreate"
      },
      "IsolationDatasetDetail": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pair_count": {
            "type": "integer",
            "title": "Pair Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "pairs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pairs"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "name",
          "pair_count",
          "created_at",
          "updated_at"
        ],
        "title": "IsolationDatasetDetail"
      },
      "IsolationDatasetSummary": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pair_count": {
            "type": "integer",
            "title": "Pair Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "name",
          "pair_count",
          "created_at",
          "updated_at"
        ],
        "title": "IsolationDatasetSummary"
      },
...
      "IsolationModelSummary": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": false
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsolationTrainMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "name",
          "created_at"
        ],
        "title": "IsolationModelSummary"
      },
      "IsolationPairUploadResult": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "uploaded": {
            "type": "integer",
            "title": "Uploaded"
          },
          "pair_indices": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pair Indices"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "uploaded",
          "pair_indices"
        ],
        "title": "IsolationPairUploadResult"
      },
      "IsolationQuotaStatus": {
        "properties": {
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          },
          "period_start": {
            "type": "string",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "title": "Period End"
          }
        },
        "type": "object",
        "required": [
          "used",
          "limit",
          "remaining",
          "period_start",
          "period_end"
        ],
        "title": "IsolationQuotaStatus"
      },
...
      "IsolationTrainRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "base_model": {
            "type": "string",
            "title": "Base Model",
            "description": "rembg-compatible base model name",
            "default": "isnet-general-use"
          },
          "epochs": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Epochs",
            "default": 20
          },
          "val_split": {
            "type": "number",
            "maximum": 0.5,
            "minimum": 0.05,
            "title": "Val Split",
            "default": 0.2
          },
          "force_min_pairs": {
            "type": "boolean",
            "title": "Force Min Pairs",
            "description": "Allow training with fewer than 20 pairs (testing only).",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "IsolationTrainRequest"
      },
      "IsolationTrainResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/IsolationTrainStatus"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsolationTrainMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "IsolationTrainResponse"
      },
      "IsolationTrainStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "IsolationTrainStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Google sign-in access token from POST /auth/google or GET /auth/google/callback."
      }
    }
  }
}